[Rename] : userPool 폴더명 변경
